% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatVIC.R
\name{MODELL_VIC}
\alias{MODELL_VIC}
\title{modell VIC}
\usage{
MODELL_VIC(VorIn, VorPa)
}
\arguments{
\item{VorIn}{please checke output of EDHMvic::InListMake_VIC()}

\item{VorPa}{please checke output of EDHMvic::PaListMake_VIC()}
}
\value{
Q(flow) of hydrostation, m^3/s
}
\description{
modell VIC
}
\examples{
ClsNa <- c("VIC", "PenmanMonteith", "GreenAmpt", "Gash", "ARNO", "G2RES")
UPMethondList = list("Shipeng", "Shipeng", "Shipeng", "Shipeng")
VICInList <- InListMake_VIC("1989-1-1",
                            "1993-12-31",
                            ClsNa,
                            MetroList,
                            GeoList,
                            GridList,
                            UPMethondList,
                            180,
                            35)
ParamterMax = c(15, 300, 600, 900, 0.9, 2.7, 0.7, 30, 9, 1,1,15,7,7,7,10,10,10,15)
ParamterMin = c(5, 50, 100, 150, 0.1, 1.3, 0.1,9, 0.1, 0.1, 0.1, 3,0.5, 0.5, 0.5 , 0.5,0.5,0.5,0.5)
ParamterModell = 0.5 * (ParamterMax + ParamterMin)
VICPaList <- PaListMake_VIC(ParamterModell)
Q <- MODELL_VIC(VICInList, VICPaList)
}
